refactor(config/s3): put s3 session out of Config struct
