+++ flexibleengine/config.go
@@ -11,15 +11,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/chnsz/golangsdk"
 	huaweisdk "github.com/chnsz/golangsdk/openstack"
 	"github.com/chnsz/golangsdk/openstack/identity/v3/domains"
-	"github.com/hashicorp/errwrap"
-	"github.com/hashicorp/go-cleanhttp"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/logging"
 	huaweiconfig "github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/helper/pathorcontents"
@@ -27,7 +21,6 @@ import (
 
 type Config struct {
 	huaweiconfig.Config
-	s3sess *session.Session
 }
 
 func (c *Config) LoadAndValidate() error {
@@ -66,7 +59,7 @@ func (c *Config) LoadAndValidate() error {
 		}
 	}
 
-	return c.newS3Session(logging.IsDebugOrHigher())
+	return nil
 }
 
 func generateTLSConfig(c *Config) (*tls.Config, error) {
@@ -258,67 +251,6 @@ func genClients(c *Config, pao, dao golangsdk.AuthOptionsProvider) error {
 	return err
 }
 
-type awsLogger struct{}
-
-func (l awsLogger) Log(args ...interface{}) {
-	tokens := make([]string, 0, len(args))
-	for _, arg := range args {
-		if token, ok := arg.(string); ok {
-			tokens = append(tokens, token)
-		}
-	}
-	log.Printf("[DEBUG] [aws-sdk-go] %s", strings.Join(tokens, " "))
-}
-
-func (c *Config) newS3Session(osDebug bool) error {
-	if c.AccessKey != "" && c.SecretKey != "" {
-		// Setup S3 client/config information for Swift S3 buckets
-		log.Println("[INFO] Building Swift S3 auth structure")
-		creds, err := GetCredentials(c)
-		if err != nil {
-			return err
-		}
-		// Call Get to check for credential provider. If nothing found, we'll get an
-		// error, and we can present it nicely to the user
-		cp, err := creds.Get()
-		if err != nil {
-			if sErr, ok := err.(awserr.Error); ok && sErr.Code() == "NoCredentialProviders" {
-				return fmt.Errorf("No valid credential sources found for S3 Provider.")
-			}
-
-			return fmt.Errorf("Error loading credentials for S3 Provider: %s", err)
-		}
-
-		log.Printf("[INFO] S3 Auth provider used: %q", cp.ProviderName)
-
-		sConfig := &aws.Config{
-			Credentials: creds,
-			Region:      aws.String(c.Region),
-			HTTPClient:  cleanhttp.DefaultClient(),
-		}
-
-		if osDebug {
-			sConfig.LogLevel = aws.LogLevel(aws.LogDebugWithHTTPBody | aws.LogDebugWithRequestRetries | aws.LogDebugWithRequestErrors)
-			sConfig.Logger = awsLogger{}
-		}
-
-		if c.Insecure {
-			transport := sConfig.HTTPClient.Transport.(*http.Transport)
-			transport.TLSClientConfig = &tls.Config{
-				InsecureSkipVerify: true,
-			}
-		}
-
-		// Set up base session for S3
-		c.s3sess, err = session.NewSession(sConfig)
-		if err != nil {
-			return errwrap.Wrapf("Error creating Swift S3 session: {{err}}", err)
-		}
-	}
-
-	return nil
-}
-
 func (c *Config) determineRegion(region string) string {
 	// If a resource-level region was not specified, and a provider-level region was set,
 	// use the provider-level region.
@@ -358,18 +290,6 @@ func (c *Config) getDomainID() (string, error) {
 	}
 
 	return all[0].ID, nil
-}
-
-func (c *Config) computeS3conn(region string) (*s3.S3, error) {
-	if c.s3sess == nil {
-		return nil, fmt.Errorf("missing credentials for Swift S3 Provider, need access_key and secret_key values for provider")
-	}
-
-	endpoint := getObsEndpoint(c, region)
-	awsS3Sess := c.s3sess.Copy(&aws.Config{Endpoint: aws.String(endpoint)})
-	s3conn := s3.New(awsS3Sess)
-
-	return s3conn, nil
 }
 
 func (c *Config) blockStorageV2Client(region string) (*golangsdk.ServiceClient, error) {
@@ -635,11 +555,4 @@ func (c *Config) sdkClient(region, serviceType string) (*golangsdk.ServiceClient
 
 func (c *Config) getHwEndpointType() golangsdk.Availability {
 	return golangsdk.AvailabilityPublic
-}
-
-func getObsEndpoint(c *Config, region string) string {
-	if endpoint, ok := c.Endpoints["oss"]; ok {
-		return endpoint
-	}
-	return fmt.Sprintf("https://oss.%s.%s/", region, c.Cloud)
 }
