+++ flexibleengine/resource_flexibleengine_s3_bucket_object_test.go
@@ -204,7 +204,7 @@ func testAccCheckS3BucketObjectVersionIdDiffers(first, second *s3.GetObjectOutpu
 
 func testAccCheckS3BucketObjectDestroy(s *terraform.State) error {
 	config := testAccProvider.Meta().(*Config)
-	s3conn, err := config.computeS3conn(OS_REGION_NAME)
+	s3conn, err := computeS3conn(config, OS_REGION_NAME)
 	if err != nil {
 		return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 	}
@@ -239,7 +239,7 @@ func testAccCheckS3BucketObjectExists(n string, obj *s3.GetObjectOutput) resourc
 		}
 
 		config := testAccProvider.Meta().(*Config)
-		s3conn, err := config.computeS3conn(OS_REGION_NAME)
+		s3conn, err := computeS3conn(config, OS_REGION_NAME)
 		if err != nil {
 			return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 		}
@@ -305,7 +305,7 @@ func testAccCheckS3BucketObjectAcl(n string, expectedPerms []string) resource.Te
 	return func(s *terraform.State) error {
 		rs, _ := s.RootModule().Resources[n]
 		config := testAccProvider.Meta().(*Config)
-		s3conn, err := config.computeS3conn(OS_REGION_NAME)
+		s3conn, err := computeS3conn(config, OS_REGION_NAME)
 		if err != nil {
 			return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 		}
@@ -393,7 +393,7 @@ func testAccCheckS3BucketObjectSSE(n, expectedSSE string) resource.TestCheckFunc
 	return func(s *terraform.State) error {
 		rs, _ := s.RootModule().Resources[n]
 		config := testAccProvider.Meta().(*Config)
-		s3conn, err := config.computeS3conn(OS_REGION_NAME)
+		s3conn, err := computeS3conn(config, OS_REGION_NAME)
 		if err != nil {
 			return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 		}
