+++ flexibleengine/resource_flexibleengine_s3_bucket_policy.go
@@ -38,7 +38,7 @@ func resourceS3BucketPolicy() *schema.Resource {
 
 func resourceS3BucketPolicyPut(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	s3conn, err := config.computeS3conn(GetRegion(d, config))
+	s3conn, err := computeS3conn(config, GetRegion(d, config))
 	if err != nil {
 		return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 	}
@@ -76,7 +76,7 @@ func resourceS3BucketPolicyPut(d *schema.ResourceData, meta interface{}) error {
 
 func resourceS3BucketPolicyRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	s3conn, err := config.computeS3conn(GetRegion(d, config))
+	s3conn, err := computeS3conn(config, GetRegion(d, config))
 	if err != nil {
 		return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 	}
@@ -99,7 +99,7 @@ func resourceS3BucketPolicyRead(d *schema.ResourceData, meta interface{}) error 
 
 func resourceS3BucketPolicyDelete(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*Config)
-	s3conn, err := config.computeS3conn(GetRegion(d, config))
+	s3conn, err := computeS3conn(config, GetRegion(d, config))
 	if err != nil {
 		return fmt.Errorf("Error creating FlexibleEngine s3 client: %s", err)
 	}
